+++ pg_documentdb/include/opclass/bson_gin_index_term.h
@@ -24,6 +24,12 @@ typedef struct BsonIndexTerm
 	/* Whether or not it's a metadata term */
 	bool isIndexTermMetadata;
 
+	/* Whether or not an undefined term is due to the
+	 * value being undefined (as opposed to the listeral
+	 * undefined).
+	 */
+	bool isValueUndefined;
+
 	/* The index term element */
 	pgbsonelement element;
 } BsonIndexTerm;
@@ -79,8 +85,12 @@ typedef struct IndexTermCreateMetadata
 } IndexTermCreateMetadata;
 
 
+bool IsSerializedIndexTermComposite(bytea *indexTermSerialized);
 bool IsSerializedIndexTermTruncated(bytea *indexTermSerialized);
 void InitializeBsonIndexTerm(bytea *indexTermSerialized, BsonIndexTerm *indexTerm);
+
+int32_t InitializeCompositeIndexTerm(bytea *indexTermSerialized, BsonIndexTerm
+									 indexTerm[INDEX_MAX_KEYS]);
 
 BsonIndexTermSerialized SerializeBsonIndexTerm(pgbsonelement *indexElement,
 											   const IndexTermCreateMetadata *
@@ -91,24 +101,18 @@ BsonCompressableIndexTermSerialized SerializeBsonIndexTermWithCompression(
 	IndexTermCreateMetadata
 	*indexMetadata);
 
-BsonIndexTermSerialized SerializeCompositeBsonIndexTerm(pgbsonelement *indexElement,
-														const IndexTermCreateMetadata *
-														indexMetadata, bool
-														hasTruncatedPaths);
+BsonIndexTermSerialized SerializeCompositeBsonIndexTerm(bytea **individualTerms, int32_t
+														numTerms);
 BsonCompressableIndexTermSerialized SerializeCompositeBsonIndexTermWithCompression(
-	pgbsonelement *indexElement,
-	const
-	IndexTermCreateMetadata
-	*
-	indexMetadata,
-	bool
-	hasTruncatedPaths);
+	bytea **individualTerms, int32_t numTerms);
 
 Datum GenerateRootTerm(const IndexTermCreateMetadata *);
 Datum GenerateRootExistsTerm(const IndexTermCreateMetadata *);
 Datum GenerateRootNonExistsTerm(const IndexTermCreateMetadata *);
 Datum GenerateRootTruncatedTerm(const IndexTermCreateMetadata *);
 Datum GenerateRootMultiKeyTerm(const IndexTermCreateMetadata *);
+Datum GenerateValueUndefinedTerm(const IndexTermCreateMetadata *termData, const
+								 char *path);
 int32_t CompareBsonIndexTerm(BsonIndexTerm *left, BsonIndexTerm *right,
 							 bool *isComparisonValid);